# swagger.yaml
openapi: 3.0.0
info:
  title: News Portal API
  description: Yangiliklar  uchun Backend API 
  version: 1.0.0

servers:
  - url: http://127.0.0.1:5000
    description: Lokal server
    

paths:
  /auth/register:
    post:
      summary: Ro'yxatdan o'tish
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: "admin"
                password:
                  type: string
                  example: "mypassword123"
      responses:
        '201':
          description: Muvaffaqiyatli ro'yxatdan o'tildi
        '400':
          description: Username band yoki malumotlar notogri
        '500':
          description: Baza bilan ulanishda xatolik
  /auth/login:
    post:
      summary: Kirish , Token olish
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:               
                  type: string
                  example: "admin"
                password:
                  type: string
                  example: "mypassword123"
      responses:
        '200':
          description: Muvaffaqiyatli kirildi (Token qaytdi)

  /api/categories:
    get:
      summary: Barcha kategoryalarni olish
      responses:
        '200':
          description: Kategoriyalar ro'yxati
    post:
      summary: Yangi kategorya yaratish
      security:
        - Bearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Texnologiya"
                slug:
                  type: string
                  example: "texnologiya"
      responses:
        '201':
          description: Kategory yaratildi


  /api/posts:
    get:
      summary: Barcha yangiliklarni olish
      parameters:
        - name: title
          in: query
          required: false
          schema:
            type: string
          example: "sun"
          description: "Title shu qiymat bilan boshlanadigan postlarni qaytaradi (prefix bo'yicha, katta-kichik harf farqsiz)"
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
          example: 1
          description: "Sahifa raqami"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
          example: 10
          description: "Bitta sahifada nechta post qaytarilishi"
      responses:
        '200':
          description: Yangiliklar roixati
    post:
      summary: Yangi yangilik qoshish
      security:
        - Bearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: "Sun'iy intellekt rivojlanishi"
                content:
                  type: string
                  example: "Yangilikning to'liq matni..."
                category_id:
                  type: integer
                  example: 1
      responses:
        '201':
          description: Yangilik yaratildi
  /api/posts/{postId}:
    get:
      summary: Bitta yangilikni (va uning commentlarini) olish
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Post ma'lumotlari, image_url, comments va comment_count bilan
        '404':
          description: Post topilmadi  
  /api/posts/{postId}/image:
    post:
      summary: Postga rasm yuklash
      security:
        - Bearer: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: Rasm muvaffaqiyatli yuklandi
        '400':
          description: Fayl notogri yoki topilmagan
        '404':
          description: Post topilmadi


  /api/pages:
    get:
      summary: Barcha sahifani olish
      responses:
        '200':
          description: Sahifa ro'yxati qaytdi
    post:
      summary: Yangi sahifa yaratish
      security:
        - Bearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - content
                - slug
              properties:
                title:
                  type: string
                  example: "Biz haqimizda"
                content:
                  type: string
                  example: "Project haqida  ma'lumot..."
                slug:
                  type: string
                  example: "biz haqimizda"
      responses:
        '201':
          description: Sahifa yaratildi
        '400':
          description: Ma'lumotlar  togri emas

  /api/comments:
    get:
      summary: Barcha izohlarni olish
      responses:
        '200':
          description: Izohlar ro'yxati qaytdi
    post:
      summary: Postga yangi izoh qo'shish
      security:              
        - Bearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - post_id
                - text
              properties:
                post_id:
                  type: integer
                  example: 1
                text:
                  type: string
                  example: "Juda ajoyib yangilik bolibdi!"
      responses:
        '201':
          description: Izoh saqlandi
        '400':
          description:  matn yoki post_id notori
        '404':
          description: post topilmadi
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT      